check: allow a merge request's created_at to override commit timestamp
Merge requests can be created at any point after a commit is pushed, so the timestamp should be a combination
of the merge request's created_at and the commit's timestamp.  For a first pass, we just use the MR's created_at
by default and allow commits with more recent timestamps to override the timestamp.

I think a better approach would be to include the commit SHA in the version, but this should work as a quick
fix (I hope).